Modified input error text margin top: styles for when input has focus. Updated capitalization of form placeholders
