+++ components/modals/EditUsernameModal.jsx
@@ -93,7 +93,7 @@ const EditUsernameModal = (props) => {
           <Form action={formAction}>
             <Form.Group className="mb-30" controlId="username">
               <Form.Label className="fw-medium">Username</Form.Label>
-              <Form.Control name="username" type="text" placeholder="username" value={username.value} onBlur={validateUsername} onChange={updateUsernameValue} className={username.valid === false && 'is-invalid'} required/>
+              <Form.Control name="username" type="text" placeholder="Enter username" value={username.value} onBlur={validateUsername} onChange={updateUsernameValue} className={username.valid === false && 'is-invalid'} required/>
               <Form.Control.Feedback type="invalid">
                 {username.message}
               </Form.Control.Feedback>
